+++ packages/code-du-travail-frontend/src/layout/PageLayout.js
@@ -3,7 +3,6 @@ import Header from "./Header.js";
 import Footer from "./Footer.js";
 
 import "@cdt/css";
-import "@reach/dialog/styles.css";
 
 const PageLayout = ({ children }) => {
   return (
