+++ packages/code-du-travail-ui/package.json
@@ -32,6 +32,7 @@
     "@babel/preset-env": "^7.3.4",
     "@babel/preset-react": "^7.0.0",
     "@cdt/css": "^2.1.0",
+    "@reach/dialog": "^0.1.2",
     "@socialgouv/eslint-config-react": "^0.5.0",
     "babel-jest": "^24.5.0",
     "docz": "v1.0.0-rc.7",
@@ -45,6 +46,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-feather": "1.1.0",
     "react-testing-library": "^6.0.0",
     "rollup": "^1.1.2",
     "rollup-plugin-babel": "^4.3.2",
