+++ packages/code-du-travail-ui/src/Container/doc.md
@@ -6,7 +6,7 @@ import "@cdt/css";
 import { Playground, Props } from "docz";
 import Container from ".";
 
-## \<Container/\>
+# Container
 
 <Playground>
   <Container>
