+++ packages/code-du-travail-ui/src/List/doc.mdx
@@ -7,7 +7,7 @@ import "@cdt/css";
 import { Playground, Props } from "docz";
 import List from ".";
 
-## \<List/\>
+# List
 
 <Playground>
   <List items={[<p>Des elements</p>,"Ou du texte"]}/>
