--- packages/code-du-travail-ui/src/icons/Like.js
@@ -1,18 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const Like = props => (
-  <svg viewBox="0 0 50 50" {...props}>
-    <title>like</title>
-    <g fill={props.fill ? props.fill : "currentColor"} fillRule="evenodd">
-      <path d="M35.833 32.423H38.7c.79 0 1.433.642 1.433 1.432 0 .79-.643 1.432-1.433 1.432h-2.867a1.433 1.433 0 1 0 0 2.865H38.7c.79 0 1.433.642 1.433 1.432 0 .79-.643 1.432-1.433 1.432h-2.867a1.433 1.433 0 1 0 0 2.864c.79 0 1.434.643 1.434 1.432 0 .79-.643 1.432-1.434 1.432H23.826c-1.39 0-2.763-.222-4.08-.661l-5.413-1.803V26.56c2.249-.427 4.45-1.85 6.436-4.178 1.94-2.275 3.513-5.29 4.207-8.067l.51-2.037h.314a2.869 2.869 0 0 1 2.867 2.865v1.43c0 1.924-.602 5.612-1.373 7.258h-2.927a1.433 1.433 0 1 0 0 2.864H38.7c.79 0 1.433.643 1.433 1.432 0 .79-.643 1.433-1.433 1.433h-2.867a1.433 1.433 0 1 0 0 2.864zm-24.366 12.89c0 .789-.643 1.431-1.434 1.431H2.867V23.83h7.166c.79 0 1.434.643 1.434 1.432v20.05zM41.902 30.99A4.279 4.279 0 0 0 43 28.126a4.303 4.303 0 0 0-4.3-4.296h-8.346c.827-2.506 1.18-5.832 1.18-7.257v-1.431A5.738 5.738 0 0 0 25.8 9.413h-1.433c-.658 0-1.231.447-1.39 1.085l-.781 3.12c-1.094 4.372-4.568 9.194-8.163 10.07a4.306 4.306 0 0 0-4-2.722h-8.6c-.79 0-1.433.641-1.433 1.432v25.779c0 .79.642 1.432 1.433 1.432h8.6a4.305 4.305 0 0 0 3.882-2.45L18.84 48.8c1.61.537 3.288.809 4.986.809h12.007c2.371 0 4.3-1.928 4.3-4.297 0-.558-.107-1.093-.302-1.583A4.305 4.305 0 0 0 43 39.584a4.279 4.279 0 0 0-1.098-2.865A4.279 4.279 0 0 0 43 33.855a4.279 4.279 0 0 0-1.098-2.864z" />
-      <path d="M7.01 41.196a1.402 1.402 0 1 1 0 2.804 1.402 1.402 0 0 1 0-2.804M28.978 5.23V2.38A1.39 1.39 0 0 0 27.576 1a1.39 1.39 0 0 0-1.402 1.38v2.85c0 .761.628 1.379 1.402 1.379a1.39 1.39 0 0 0 1.402-1.38M21.954 9.867a1.643 1.643 0 0 0 0-2.324L19.63 5.22a1.643 1.643 0 1 0-2.323 2.323l2.323 2.324a1.643 1.643 0 0 0 2.324 0M35.522 9.867l2.323-2.323a1.643 1.643 0 1 0-2.323-2.324l-2.323 2.323a1.643 1.643 0 0 0 2.323 2.324" />
-    </g>
-  </svg>
-);
-
-export default Like;
-
-Like.propTypes = {
-  fill: PropTypes.string
-};
