--- packages/code-du-travail-ui/src/icons/Unlike.js
@@ -1,18 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const Unlike = props => (
-  <svg viewBox="0 0 50 50" width="1em" height="1em" {...props}>
-    <title>unlike</title>
-    <g fill={props.fill ? props.fill : "currentColor"} fillRule="evenodd">
-      <path d="M7.167 18.577H4.3c-.79 0-1.433-.642-1.433-1.432 0-.79.643-1.432 1.433-1.432h2.867a1.433 1.433 0 1 0 0-2.865H4.3c-.79 0-1.433-.642-1.433-1.432 0-.79.643-1.432 1.433-1.432h2.867a1.433 1.433 0 1 0 0-2.864c-.79 0-1.434-.643-1.434-1.432 0-.79.643-1.432 1.434-1.432h12.007c1.39 0 2.763.222 4.08.661l5.413 1.803V24.44c-2.249.427-4.45 1.85-6.436 4.178-1.94 2.275-3.513 5.29-4.207 8.067l-.51 2.037H17.2a2.869 2.869 0 0 1-2.867-2.865v-1.43c0-1.924.602-5.612 1.373-7.258h2.927a1.433 1.433 0 1 0 0-2.864H4.3c-.79 0-1.433-.643-1.433-1.432 0-.79.643-1.433 1.433-1.433h2.867a1.433 1.433 0 1 0 0-2.864zm24.366-12.89c0-.789.643-1.431 1.434-1.431h7.166V27.17h-7.166c-.79 0-1.434-.643-1.434-1.432V5.688zM1.098 20.01A4.279 4.279 0 0 0 0 22.874a4.303 4.303 0 0 0 4.3 4.296h8.346c-.827 2.506-1.18 5.832-1.18 7.257v1.431a5.738 5.738 0 0 0 5.734 5.729h1.433c.658 0 1.231-.447 1.39-1.085l.781-3.12c1.094-4.372 4.568-9.194 8.163-10.07a4.306 4.306 0 0 0 4 2.722h8.6c.79 0 1.433-.641 1.433-1.432V2.823c0-.79-.642-1.432-1.433-1.432h-8.6a4.305 4.305 0 0 0-3.882 2.45L24.16 2.2a15.744 15.744 0 0 0-4.986-.809H7.167c-2.371 0-4.3 1.928-4.3 4.297 0 .558.107 1.093.302 1.583A4.305 4.305 0 0 0 0 11.416c0 1.1.416 2.104 1.098 2.865A4.279 4.279 0 0 0 0 17.145c0 1.099.416 2.103 1.098 2.864z" />
-      <path d="M35.99 9.804a1.402 1.402 0 1 1 0-2.804 1.402 1.402 0 0 1 0 2.804" />
-    </g>
-  </svg>
-);
-
-export default Unlike;
-
-Unlike.propTypes = {
-  fill: PropTypes.string
-};
