+++ packages/code-du-travail-ui/src/icons/index.js
@@ -1,5 +1,2 @@
-export { default as Like } from "./Like";
-export { default as Unlike } from "./Unlike";
 export { default as Question } from "./Question";
 export { default as Lost } from "./Lost";
-export { default as Search } from "./Search";
